+++ tests/test_io.py
@@ -3,7 +3,8 @@ import pandas as pd
 import numpy as np
 import pytest
 
-from pyam import IamDataFrame
+from pyam import IamDataFrame, read_datapackage
+from pyam.testing import assert_iamframe_equal
 
 from conftest import TEST_DATA_DIR
 
@@ -41,9 +42,8 @@ def test_io_xlsx(test_df, meta_args):
         # read from xlsx
         import_df = IamDataFrame(file, **meta_args[1])
 
-        # assert that `data` and `meta` tables are equal and delete file
-        pd.testing.assert_frame_equal(test_df.data, import_df.data)
-        pd.testing.assert_frame_equal(test_df.meta, import_df.meta)
+        # assert that IamDataFrame instances are equal and delete file
+        assert_iamframe_equal(test_df, import_df)
         os.remove(file)
 
 
@@ -74,3 +74,20 @@ def test_load_rcp_database_downloaded_file(test_df_year):
         TEST_DATA_DIR, 'test_RCP_database_raw_download.xlsx')
     )
     pd.testing.assert_frame_equal(obs_df.as_pandas(), exp)
+
+
+def test_io_datapackage(test_df):
+    file = 'foo.zip'
+
+    # add column to `meta`
+    test_df.set_meta(['a', 'b'], 'string')
+
+    # write to datapackage
+    test_df.to_datapackage(file)
+
+    # read from csv
+    import_df = read_datapackage(file)
+
+    # assert that IamDataFrame instances are equal and delete file
+    assert_iamframe_equal(test_df, import_df)
+    os.remove(file)